backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A description of your project"
authors = ["Nicole Sorokin <nicole.sorokin@arctiq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.95.0"
uvicorn = { version = "^0.23.2", extras = ["standard"] }
google-cloud-storage = "^2.8.0"
openai = "^0.28.0"
pyaudio = "^0.2.11"  # Adjust version as necessary
pydub = "^0.25.1" 
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
